Add navigation links and improve styles
